requirements.txt
faiss-cpu
fastembed
rank-bm25
fastapi
uvicorn
pypdf
sqlalchemy
cryptography